alpha

+++ MapCollator/App.xaml.cs
@@ -7,7 +7,6 @@ using System.Text;
 using System.Threading.Tasks;
 using System.Windows;
 using System.IO;
-using System.IO.Compression;
 using System.Diagnostics;
 
 
@@ -15,26 +14,44 @@ namespace MapCollator
 {
     public class MutiThreading
     {
+        public static int threadCount = Environment.ProcessorCount; 
+        public static List<System.Threading.Thread> threads = new List<System.Threading.Thread>();
+
+        public static Stopwatch sw = new Stopwatch(); //方便调试，查看耗时
+
+
+        public static void MakeThreads()
+        {
+            Console.WriteLine(threadCount);//控制台下打印处理器逻辑核心数量
+
+            for (int i = 0; i < threadCount; i++)
+            {
+                threads.Add(new System.Threading.Thread(App.distribute));
+            }
+
+        }
         public static void StartMutiThreading()
         {
-            System.Threading.Thread t1 = new System.Threading.Thread(App.distribute);
-            System.Threading.Thread t2 = new System.Threading.Thread(App.distribute);
-            System.Threading.Thread t3 = new System.Threading.Thread(App.distribute);
-            System.Threading.Thread t4 = new System.Threading.Thread(App.distribute);
-            t1.Start();
-            t2.Start();
-            t3.Start();
-            t4.Start();
+            sw.Start();
 
-            t1.Join();
-            t2.Join();
-            t3.Join();
-            t4.Join();
+            for (int i = 0; i < threadCount; i++)
+            {
+                threads[i].Start();
+            }
+            foreach (var thread in threads)
+            {
+                thread.Join();
+            }
 
-            t1.Abort();
-            t2.Abort();
-            t3.Abort();
-            t4.Abort();
+            sw.Stop();
+            Console.WriteLine(sw.ElapsedMilliseconds.ToString());//方便调试，查看耗时
+
+            threads.Clear();
+            for (int i = 0; i < threadCount; i++)
+            {
+                threads.Add(new System.Threading.Thread(App.distribute));
+            }
+
         }
     }
     
@@ -42,14 +59,12 @@ namespace MapCollator
     {
         public static string path;
     }
-    /// <summary>
-    /// App.xaml 的交互逻辑
-    /// </summary>
     public partial class App : Application
     {
         public static string path, folderName, packName, artists, creator, HP, OD;
-        private static Dictionary<string, bool> allFileDict = new Dictionary<string, bool>();
-        private static List<string> allFileList = IO.allFileList;
+        public static Dictionary<string, bool> allFileDict = new Dictionary<string, bool>();
+        public static List<string> allFileList = IO.allFileList;
+        private static readonly object _locker = new object();
 
         public class Program
         {
@@ -62,7 +77,6 @@ namespace MapCollator
 
             public static void Start()
             {
-
                 //创建新文件夹存放生成的新文件
                 folderName = Path.Combine(Path.Combine(path.Replace(Path.GetFileName(path), ""), String.Format("{0}{1}{2}", artists, " - ", packName)));
                 Directory.CreateDirectory(folderName);
@@ -91,21 +105,14 @@ namespace MapCollator
                     IO.PackToOsz(folderName);
                     //删除临时文件夹
                     Directory.Delete(Path.Combine(AppDomain.CurrentDomain.BaseDirectory, folderName), true);
-                    MessageBox.Show("Packaged successfully!");
-                    allFileDict.Clear();
-                    allFileList.Clear();
-                    StructuralAnalysis.mainDict.Clear();
-                    StructuralAnalysis.opt.Clear();
-
-
-        }
-    }
+                }
+            }
         }
         //循环读取每个osu文件以及修改
         public static void distribute()
         {
             //线程同步
-            lock(allFileDict)
+            lock(_locker)
             {
                 foreach (string item in allFileList)
                 {
@@ -114,10 +121,9 @@ namespace MapCollator
                         if (allFileDict[item] == false)
                         {
                             IO.path = item;
-
                             IO.ReadFile();
-                            allFileDict[item] = true;
 
+                            allFileDict[item] = true;
                         }
                     }
                 }
@@ -140,8 +146,8 @@ namespace MapCollator
         static string opt_a, opt_b;
         public static void ReadFile()
         {
-
             pathWithOutFileName = path.Replace(Path.GetFileName(path), "");
+            //以下using似乎比较多余，然而懒得优化了
             using (StreamReader f1 = new StreamReader(path))
             {
                 while ((line = f1.ReadLine()) != null)
@@ -149,12 +155,21 @@ namespace MapCollator
                     if (line.Contains("Title:"))
                     {
                         title = line.Replace("Title:", "");
+                        if (title == "")
+                        {
+                            title = "Undefined";
+                        }
                     }
                     if (line.Contains("Artist:"))
                     {
                         artist = line.Replace("Artist:", "");
+                        if (artist == "")
+                        {
+                            artist = "Undefined";
+                        }
                     }
                     if (line.Contains("Version:"))
+
                     {
                         version = line.Replace("Version:", "");
                         newVersion = String.Format("{1}{3}{0}{4}{2}{5}", title, artist, version, " - ", " [", "] ");
@@ -209,28 +224,32 @@ namespace MapCollator
                             audioFileName = line.Replace("AudioFilename: ", "");
                             audioFileFormat = Path.GetExtension(audioFileName);
                             //新的音频文件名字
-                            line = FilterIllegalChars_2(String.Format("{4}{0}{1}{2}{5}{3}", artist, " - ", title, audioFileFormat, "AudioFilename: ", opt_a));
-                            newAudioFileName = String.Format("{0}{1}{2}{4}{3}", artist, " - ", title, audioFileFormat, opt_a);
-                            newAudioFileName = FilterIllegalChars_2(newAudioFileName);
+                            newAudioFileName = String.Format("{0}{1}{2}{4}{3}", artist.Replace(":", ""), " - ", title.Replace(":", ""), audioFileFormat, opt_a);
+                            newAudioFileName = FilterIllegalChars_2(newAudioFileName.Replace(":", ""));
+                            line = FilterIllegalChars_2(String.Format("{1}{0}", newAudioFileName, "AudioFilename: "));
                         }
-                        if (line.Contains("0,0,\"") && line.Contains("\",0,0"))
+                        if (line.Contains("0,0,\"") && (line.Contains(".jpg") || line.Contains(".png")))
                         {
-                            backgroundFileName = line.Replace("0,0,\"", "").Replace("\",0,0", "");
+                            backgroundFileName = line.Replace("0,0,\"", "").Replace("\",0,0", "").Replace("\",0,80", "").Replace("\"", "");
                             backgroundFileFormat = Path.GetExtension(backgroundFileName);
                             //新的bg文件名字
-                            line = String.Format("{4}{0}{1}{2}{6}{3}{5}", FilterIllegalChars_2(artist), " - ", FilterIllegalChars_2(title.TrimEnd('.')), backgroundFileFormat, "0,0,\"", "\",0,0", opt_b);
-                            newBgFileName = String.Format("{0}{1}{2}{4}{3}", artist, " - ", title, backgroundFileFormat, opt_b);
-                            newBgFileName = FilterIllegalChars_2(newBgFileName);
+                            newBgFileName = String.Format("{0}{1}{2}{4}{3}", artist.Replace(":", ""), " - ", title.Replace(":", ""), backgroundFileFormat, opt_b);
+                            newBgFileName = FilterIllegalChars_2(newBgFileName.Replace(":", ""));
+                            line = String.Format("{1}{0}{2}", newBgFileName,"0,0,\"", "\",0,0");
                         }
                         if (line.Contains("Version:"))
                         {
-                            version = line.Replace("Version:", "");
+                            version = FilterIllegalChars_2(line.Replace("Version:", ""));
                             newVersion = String.Format("{1}{3}{0}{4}{2}{5}", title, artist, version, " - ", " [", "] ");
                             line = String.Format("{0}{1}", "Version:", newVersion);
                         }
                         if (line.Contains("Title:"))
                         {
                             title = line.Replace("Title:", "");
+                            if (title == "")
+                            {
+                                title = "Undefined";
+                            }
                             line = String.Format("{0}{1}", "Title:", packName);
                         }
                         if (line.Contains("TitleUnicode:"))
@@ -240,6 +259,10 @@ namespace MapCollator
                         if (line.Contains("Artist:"))
                         {
                             artist = line.Replace("Artist:", "");
+                            if (artist == "")
+                            {
+                                artist = "Undefined";
+                            }
                             line = String.Format("{0}{1}", "Artist:", artists);
                         }
                         if (line.Contains("ArtistUnicode:"))
@@ -268,33 +291,47 @@ namespace MapCollator
                         }
                         //这里用来写入修改后的内容
                         fw.WriteLine(line);
-
                     }
                 }
             }
-            try
+            //复制所关联的音频和bg文件到新的文件夹并应用新的名字
+            if (backgroundFileName != null)
             {
-                //复制所关联的音频和bg文件到新的文件夹并应用新的名字
                 oldAudioFilePath = Path.Combine(pathWithOutFileName, audioFileName);
                 oldBgFilePath = Path.Combine(pathWithOutFileName, backgroundFileName);
                 newAudioFilePath = Path.Combine(folderName, newAudioFileName);
                 newBgFilePath = Path.Combine(folderName, newBgFileName);
-                File.Copy(oldAudioFilePath, newAudioFilePath);
-                File.Copy(oldBgFilePath, newBgFilePath);
+                if (File.Exists(oldAudioFilePath))
+                {
+                    if (File.Exists(newAudioFilePath) == false)
+                    {
+                        File.Copy(oldAudioFilePath, newAudioFilePath);
+                    }
+                }
+                if (File.Exists(oldBgFilePath))
+                {
+                    if (File.Exists(newBgFilePath) == false)
+                    {
+                        File.Copy(oldBgFilePath, newBgFilePath);
+                    }
+                }
             }
-            catch (FileNotFoundException)
+            else
             {
+                oldAudioFilePath = Path.Combine(pathWithOutFileName, audioFileName);
+                newAudioFilePath = Path.Combine(folderName, newAudioFileName);
+                if (File.Exists(oldAudioFilePath))
+                {
+                    if (File.Exists(newAudioFilePath) == false)
+                    {
+                        File.Copy(oldAudioFilePath, newAudioFilePath);
+                    }
+                }
             }
-            catch (IOException)
-            {
-            }
-            catch (ArgumentNullException)
-            {
-            }
+
             title = null; artist = null; version = null; audioFileName = null; backgroundFileName = null;
             newVersion = null; newAudioFileName = null; newBgFileName = null;
             newFilePath = null;
-
         }
     public static string[] fileList;
         public static List<string> allFileList = new List<string>();
@@ -333,46 +370,40 @@ namespace MapCollator
             catch (ArgumentException)
             {
             }
-            str = str.Replace("\\", "_");
-            str = str.Replace("/", "_");
-            str = str.Replace("*", "_");
-            str = str.Replace("?", "_");
-            str = str.Replace("\"", "_");
-            str = str.Replace("<", "_");
-            str = str.Replace(">", "_");
-            str = str.Replace("|", "_");
-            str = str.Replace(":", "_");
-            str = str.Replace("..", ".");
-
+            str = str.Replace("\\", "");
+            str = str.Replace("/", "");
+            str = str.Replace("*", "");
+            str = str.Replace("?", "");
+            str = str.Replace("\"", "");
+            str = str.Replace("<", "");
+            str = str.Replace(">", "");
+            str = str.Replace("|", "");
             return str;
 
         }
         public static string FilterIllegalChars_2(string path)
         {
             string str = path;
-            str = str.Replace("\\", "_");
-            str = str.Replace("/", "_");
-            str = str.Replace("*", "_");
-            str = str.Replace("?", "_");
-            str = str.Replace("\"", "_");
-            str = str.Replace("<", "_");
-            str = str.Replace(">", "_");
-            str = str.Replace("|", "_");
-            str = str.Replace(":", "_");
-            str = str.Replace("..", ".");
+            str = str.Replace("\\", "");
+            str = str.Replace("/", "");
+            str = str.Replace("*", "");
+            str = str.Replace("?", "");
+            str = str.Replace("\"", "");
+            str = str.Replace("<", "");
+            str = str.Replace(">", "");
+            str = str.Replace("|", "");
             return str;
 
         }
         public static void PackToOsz(string rootPath)
         {
-            string Args = "-tZip a \"" + rootPath + ".osz\" \"" + rootPath + "\\*\" -mx3";
+            string Args = "-tZip a \"" + rootPath + ".osz\" \"" + rootPath + "\\*\" -mx2";
             Process pro = new Process();
             pro.StartInfo.FileName = @"Ref\7z.exe";
             pro.StartInfo.Arguments = Args;
             pro.StartInfo.WindowStyle = ProcessWindowStyle.Hidden;
             pro.Start();
             pro.WaitForExit();
-
         }
     }
     //预先解析文件结构方便下一行注释的内容
@@ -426,14 +457,15 @@ namespace MapCollator
                             if (backgroundFileName != null)
                             {
                                 break;
-                                
                             }
                         }
                         List<string> osuFile = new List<string>() { title, artist, version, newVersion, audioFileName, backgroundFileName };
-                        mainDict.Add(path, osuFile);
+                        if (mainDict.ContainsKey(path) == false)
+                        {
+                            mainDict.Add(path, osuFile);
+                        }
                         backgroundFileName = null;
                     }
-
                 }
             }
             //有关判断是否一首歌是否有不同音频和背景的逻辑
@@ -467,10 +499,9 @@ namespace MapCollator
                                 Optional = "_" + mainDict[path][2];
                                 opt.Add(path, new List<string> { "bg", Optional });
                             }
-                        }   
+                        }
                     }
                 }
-
             }
         }
     }
